fix(ralplan): keep state-runtime test shard native-free; add changelog entry
- Move native-dependent WI-6 HUD tests (artifact/state ordering, lane-verdict
  precedence) from state-runtime.test.ts to ralplan-runtime.test.ts so the
  gjc-state-gates runtime CI shard stays native-free
- Add Unreleased changelog entry for gjc.ralplan.maxReviewPassesPerLane,
  per-lane review budget / lane PLANNING-STUCK, --lane-verdict HUD carriage,
  and the reviewer re-review ratchet contracts
